+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unpublished]
 ### Fixed
 - Fixed `~statues` command calculating for 5 statues instead of 6.
+- Fixed mod commands being used while mentioning bot.
 
 ## [2.3.4] - 2016-07-25
 ### Changed